fix(tavily): strip unsupported schema mechanism
