+++ api/server/index.js
@@ -2,7 +2,6 @@ require('dotenv').config();
 const fs = require('fs');
 const path = require('path');
 require('module-alias')({ base: path.resolve(__dirname, '..') });
-require('./utils/patchZodToJsonSchema');
 const cors = require('cors');
 const axios = require('axios');
 const express = require('express');
